Add compatibility for Redmine 2.x, sort of using this as a guide: http://www.redmine.org/boards/3/topics/31445#Relative-url-root
Unfortunately, Redmine 1.x cannot compile a block statement with <%= (which is used by Rails 3), so the clearest way to handle this was to have a separate file for each version of Redmine (the files differ by a single character :/ )

The label in the "my accounts" page seems to be working a bit differently in 2.x. Instead of displaying the text specified by the plugin's language file (as 1.x does), 2.x must be using the text from Redmine's file. I would consider this a bug in Redmine.